+++ packages/system/src/index.ts
@@ -1,8 +1,8 @@
 export * from "./omit"
-export * from "./useMergeValue"
+export * from "./use-merge-value"
 export * from "./is"
 export * from "./raf"
-export * from "./throttleByRaf"
+export * from "./throttle-by-raf"
 export * from "./keycode"
 export * from "./easing"
 export * from "./dayjs"
