+++ packages/system/src/style.ts
@@ -2,6 +2,10 @@ interface StyleObj extends CSSStyleDeclaration {
   [index: string]: any
 }
 
+/**
+ * Get element style value, e.g. width, margin-left..
+ *
+ */
 export function getStyle(
   element: HTMLElement,
   prop: string,
